LTM 1.6 updates (#156)
* LTM 1.6 updates

This adds the fix for docs, backported from 2.3.1. This also fixes the CI so that it runs with Nautobot 1.6 instead of 2.3.x (Stable). Changed CI minimum version to hardcoded 1.6.0 as that is all that the LTM is meant to support, so we don't need to worry about trying to test anything older.

* Update nautobot-app/{{ cookiecutter.project_slug }}/.github/workflows/ci.yml

Co-authored-by: Gary Snider <75227981+gsnider2195@users.noreply.github.com>

* Update nautobot-app/{{ cookiecutter.project_slug }}/pyproject.toml

Co-authored-by: Gary Snider <75227981+gsnider2195@users.noreply.github.com>

* Apply peer suggestions

---------

Co-authored-by: Gary Snider <75227981+gsnider2195@users.noreply.github.com>