Drift Manager Improvements (#165)
* Update towncrier_template.j2

If a change fragement does not have an issue tied to it (for instance the new drift manager +nautobot-app-v2.3.2.housekeeping) it will render as:

- Rebaked from the cookie `Nautobot-app-v2.3.2`.

Instead of the previous:

- - Rebaked from the cookie `Nautobot-app-v2.3.2`.

* Update tag-nautobot-app-v2.yml

Add Ruff and Poetry lock to post actions, remove black.  This also sets the PR to Ready for Review, vs opening in Draft.