Make min/max nautobot version a template global instead of input
This change removes the min_nautobot_version and max_nautobot_version and replaced them with a Jinja2 global. This change allows the nautobot versions to be set in the template instead of by the users.

This is especially important with Drift Manager as it will be unable to update the version unless the .cookiecutter.json was updated before running drift-manager.